requirements.txt
rpyc
pypresence